Merge Latest

+++ src/client/command/commands/gm0/CheckDpsCommand.java
@@ -1,19 +1,16 @@
 /*
     This file is part of the HeavenMS MapleStory Server, commands OdinMS-based
     Copyleft (L) 2016 - 2018 RonanLana
-
     This program is free software: you can redistribute it and/or modify
     it under the terms of the GNU Affero General Public License as
     published by the Free Software Foundation version 3 as published by
     the Free Software Foundation. You may not use, modify or distribute
     this program under any other version of the GNU Affero General Public
     License.
-
     This program is distributed in the hope that it will be useful,
     but WITHOUT ANY WARRANTY; without even the implied warranty of
     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
     GNU Affero General Public License for more details.
-
     You should have received a copy of the GNU Affero General Public License
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
@@ -94,7 +91,7 @@ public class CheckDpsCommand extends Command {
 
         monster.addListener(new MonsterListener() {
             @Override
-            public void monsterDamaged(MapleCharacter from, int trueDmg) {
+            public void monsterDamaged(MapleCharacter from, long trueDmg) {
                 if (from.getId() == player.getId()) {
                     if (player.getDpsStart() == -1) {
                         player.setDpsStart(System.currentTimeMillis());
@@ -106,7 +103,7 @@ public class CheckDpsCommand extends Command {
             @Override
             public void monsterKilled(int aniTime) {}
             @Override
-            public void monsterHealed(int trueHeal) {}
+            public void monsterHealed(long trueHeal) {}
         });
 
         player.getMap().spawnMonsterOnGroundBelow(monster, player.getPosition());
