+++ src/server/life/MonsterListener.java
@@ -4,6 +4,6 @@ import client.MapleCharacter;
 public interface MonsterListener {
     
     public void monsterKilled(int aniTime);
-    public void monsterDamaged(MapleCharacter from, int trueDmg);
-    public void monsterHealed(int trueHeal);
+    public void monsterDamaged(MapleCharacter from, long trueDmg);
+    public void monsterHealed(long trueHeal);
 }
